basic tic tac toe game
Created the tic tac toe game with score and timer. Still need to make it so it will officially stop the game and show the scoreboard instead of the game still technically going after the timer goes off. Later will add AI to go up against with a customizable board.